KRW 페이지 숫자 정리

+++ app/src/main/java/com/example/upbitsampleapp/entities/CoinData.kt
@@ -1,9 +1,11 @@
 package com.example.upbitsampleapp.entities
 
+import java.math.BigDecimal
+
 data class CoinData(
     val korName: String,
     val engName: String,
-    val currentPrice: Int,
+    val currentPrice: BigDecimal,
     val perYesterday: Double,
-    val tradePrice: Double
+    val tradePrice: BigDecimal
 )
